+++ kcache.h
@@ -108,7 +108,7 @@ double lasvm_kcache_query(lasvm_kcache_t *self, int i, int j);
    and vice-versa.
 */
 
-float *lasvm_kcache_query_row(lasvm_kcache_t *self, int i, int len);
+double *lasvm_kcache_query_row(lasvm_kcache_t *self, int i, int len);
 
 /* --- lasvm_kcache_status_row
    Returns the number of cached entries for row i.
